CDK virtual scroll documentation issue fix
There's a discrepancy in the documentation sample of ng-scrollbar combined with cdk-virtual-scroll-viewport  and stackbiz sample. Because of misleading documentation I could not configure properly the ng-scrollbar to work with virtual scroll. Apart from that, the stackbiz link was leading to the production version without possibility to see the source code. Both of the problems are addressed in this proposed changes.